Boggle 1.0
This is the version wherein the basic GUI of the grid is ready and a
random board appears when you press the boggle button. The problem that
I could figure out here is that I should actually add more vowels so
that more number of words can be formed in the board. I am currently
working on that.